flaskbook_api/requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.2
coverage==7.13.5
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
dnspython==2.8.0
email-validator==2.3.0
filelock==3.28.0
Flask==3.1.3
Flask-DebugToolbar==0.16.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fsspec==2026.3.0
greenlet==3.4.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.11
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
opencv-python==4.13.0.92
packaging==26.1
pillow==12.2.0
pluggy==1.6.0
Pygments==2.20.0
PyMySQL==1.1.2
pytest==9.0.3
pytest-cov==7.1.0
python-dotenv==1.2.2
setuptools==81.0.0
SQLAlchemy==2.0.49
sympy==1.14.0
torch==2.11.0
torchaudio==2.11.0+cpu
torchvision==0.26.0
triton==3.6.0
typing_extensions==4.15.0
Werkzeug==3.1.8
WTForms==3.2.1
